Fixes possible wrong encoding with 32bit on Radius

+++ include/ntop_typedefs.h
@@ -1217,11 +1217,13 @@ typedef struct {
 } InOutTraffic;
 
 typedef struct {
-  u_int64_t bytes_sent;
-  u_int64_t bytes_rcvd;
-  u_int64_t packets_sent;
-  u_int64_t packets_rcvd;
-  u_int8_t terminate_cause;
+  /* All int32 because following the rfc, they use 32 bits */
+  /* See: https://datatracker.ietf.org/doc/html/rfc2866 */
+  u_int32_t bytes_sent;
+  u_int32_t bytes_rcvd;
+  u_int32_t packets_sent;
+  u_int32_t packets_rcvd;
+  u_int32_t terminate_cause;
 } RadiusTraffic;
 
 typedef enum {
